requirements.txt
Django==1.6.2
Pillow==2.3.1
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
django-colorful==0.1.3
djangorestframework==2.3.13
django-pylibmc==0.5.0
pylibmc==1.2.3
pytz==2014.1
psycopg2==2.5.2
uWSGI==2.0.3
wsgiref==0.1.2
-e git+https://github.com/jsocol/django-ratelimit.git@fab72f90dd88149c39d2fe84df2a1b6c3439a442#egg=django-ratelimit
